Solved a warning.. and realized that student form doesn't need to be a container.
"Warning: Failed prop type: Invalid prop `component` of type `object` supplied to `Route`, expected `function`."
